requirements.txt
# 基础依赖
pillow>=9.0.0
numpy>=1.20.0
requests>=2.25.0
pyyaml>=6.0
python-dateutil>=2.8.0

# 跨平台依赖
pyautogui>=0.9.53
pygetwindow>=0.0.9
pyperclip>=1.8.2

# Windows特定依赖
pywin32>=300; sys_platform == 'win32'

# Mac特定依赖
# 无需额外依赖，使用系统命令

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
